:wrench: <Init/Menu/Shutdown> refactor, part 2/3
MainThread (file + class) renamed to MainMenu
